Formatting improvements

+++ procs/adapter.py
@@ -18,25 +18,23 @@ SHUFFLE_BUFFER_SIZE = 1000
 
 # testing function to enhance input image
 # should be rewritten to tf
-def hist_equalization_color(img,  clipLimit = 2.0, tileGridSize=(8,8)):
+def hist_equalization_color(img, clipLimit = 2.0, tileGridSize=(8,8)):
         
     # to HSV format
     hsv_img = opencv.cvtColor(img, opencv.COLOR_BGR2HSV)
         
-    h, s, v = hsv_img[:,:,0], hsv_img[:,:,1], hsv_img[:,:,2]
+    h, s, v = hsv_img[:, :, 0], hsv_img[:, :, 1], hsv_img[:, :, 2]
     # create a CLAHE object (Arguments are optional).
     # clipLimit	= Threshold for contrast limiting. 
     # apply CLAHE on V part
     clahe = opencv.createCLAHE(clipLimit=clipLimit, tileGridSize=tileGridSize)
     v = clahe.apply(v)
     # complete image
-    hsv_img = np.dstack((h,s,v))
+    hsv_img = np.dstack((h, s, v))
     # return back to BGR
     print('Clahe')
-    return  opencv.cvtColor(hsv_img, opencv.COLOR_HSV2BGR)
+    return opencv.cvtColor(hsv_img, opencv.COLOR_HSV2BGR)
     
-    
-
 
 def read_mask(filename):
 
@@ -76,7 +74,7 @@ def aug_random_brightness(img, mask):
 
 class DataAdapter(object):
 
-    def __init__(self, fn_csv: str, patch_size: int = 128, patch_overlap_ratio: int = 0.0, test_ratio: float = 0.2, augmented_ratio: float = .3, enhancement: bool = False):
+    def __init__(self, fn_csv: str, patch_size: int = 128, patch_overlap_ratio: float = 0.0, test_ratio: float = 0.2, augmented_ratio: float = .3, enhancement: bool = False):
 
         self._df_paths = None
